Make warnings checker a generic IrisTest function.
